Support new style external networks and multiple external networks
Switch the generated configuration to use "new" style external networks when
ext-port is not set.  In this case we configure external_network_bridge =
(intentionally blank), gateway_external_network_id = (blank) and update the
README with information on using this new style of configuration.

The current template configures external networks by using the default
external_network_bridge=br-ex (implied when not set).  This activates legacy
code which assumes that a single external network exists on that bridge and the
L3 Agent directly plugs itself in.  provider:network_type,
provider:physical_network and provider:segmentation_id are ignored.  You cannot
create multiple networks and you cannot use segmented networks (e.g. VLAN)

By setting external_network_bridge = (intentionally blank) the L2 Agent handles
the configuration instead, this allows us to create multiple networks and also
to use more complex network configurations such as VLAN.  It is also possible
to use the same physical connection with different segmentation IDs for both
internal and external networks, as well as multiple external networks.

Legacy/existing configurations where ext-port is set generate the same
configuration as previous and should continue to work as before.  I do not
believe it to be easy to migrate existing setups to the "new" style
configuration automatically as changes to the neutron network configuration may
be required (specifically: provider:physical_network will now be used when it
was not before, and may not be correct) and the physical port needs to be moved
from br-ex to br-data which the charm does not currently handle and is likely
to error as it does not attempt removal first.  Further work may be possible in
this area.

For information about this new style of configuration being preferred, see
discussions in LP#1491668, LP#1525059 and
http://docs.openstack.org/liberty/networking-guide/scenario-classic-ovs.html

Change-Id: I8d2bb8098e080969e0445293b1ed79714b2c964f
Related-Bug: #1491668
Related-Bug: #1525059